Changed most variables to be fixed size variants

+++ game.h
@@ -1,14 +1,13 @@
 #ifndef TERMGAME_H
 #define TERMGAME_H
 
-#include <unistd.h>
-#include <pthread.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 #define SCREEN_FD STDOUT_FILENO
 #define FRAMETIME_GOAL 16666666
 
-void game_init(int level_select);
+void game_init(uint8_t level_select);
 bool game_loop();
 void game_shutdown();
 void draw_gamescreen();
